Major Update for Product.
Added the xmlsitemap node settings for both the Category and Product.

Updated the views to be separated by type node, list, admin.

Added category for panel panes to be node, list, admin.

Updated the node level to have the secondary images pagered so that only four images are shown at a time.

Added node permisions.

Signed-off-by: Staris <jaison@worxco.com>